Refactor: modificar o interface menu para menu-lateral.
