+++ Elogica-RH-Frontend/src/app/shared/interfaces/menu-lateral.ts
